Modified Review Cross Details
The cross column has been removed from the review cross details panel.

We have added in cross female gid and cross male gid columns instead. It was agreed that this was more helpful for breeders to identify the resultant crosses.

Additionally female and male parent are populated with the preferred name.

issue: BMS-3080
reviewer: none